chore: update bifrost/core to v1.1.13 and remove local replace (#205)
This PR updates the `github.com/maximhq/bifrost/core` dependency from v1.1.12 to v1.1.13 in the transports module and removes the local replacement directive that was pointing to the local `../core` directory. This change allows the transports module to use the published v1.1.13 version of the core module instead of the local development version.